Revert "This is the Sixth optimisation to the code."
This reverts commit 443736eef489451c44e31a34d745c06ae6e01d31.

Conflicts:

	lattice.cpp
	sim-state.cpp

This was done due to a bug that was introduced by Sixth optimisation (returning references instead of pointers
for Lattice::getN() and Lattice::setN()). Although the bug was fixed it will cause more problems when we try to
introduce a needed modification (not breaking out the loop when encountering a nanoparticle cell) in sim-state.